+++ src/components/List/listChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "New Feature",
+    affects: ["Styles"],
+    notes: [
+      "Adds `showRowDividers` prop to show/hide dividers between rows only for the description list variant.",
+    ],
+  },
+  {
     date: "2024-03-28",
     version: "3.0.1",
     type: "Update",